Reorder exercises by difficulty
