+++ exercises/solutions/solution_07.py
@@ -1,65 +1,81 @@
-# Calendar buttons (hit month and expect it in the date field)
-def test_hitting_calendar_buttons_is_reflected_by_date_field(page):
-    # 1. On the Kiwi.com website hit the Departure date field
+# Sorting panel actions can be used for sorting the search results
+# TODO: run without any --slowmo (slow execution leads to some weird loading on steps 2 and 3; or fix it)
+def test_sorting_panel_actions_can_be_used_for_sorting_the_search_results(page):
+    # 1. Steps 1-9. from the Searching for a connection displays results scenario, but with the From location
+    # set to Brno and the To location set to Bucharest
     # 1.1. Open the kiwi.com website (wait for page to load)
     page.goto("https://www.kiwi.com/en/")
     page.click("[data-test='CookiesPopup-Accept']")
     assert page.is_visible("text=Book cheap flights other sites simply can’t find.")
 
-    # 1.2. Hit the Departure date field
-    page.click("[data-test=SearchDateInput]:has-text('Departure')")
-    page.wait_for_selector("[data-test=NewDatePickerOpen]", state="visible")
+    # 1.2. Clear the `from` location
+    page.click("[data-test=PlacePickerInputPlace-close]")
+    page.wait_for_selector("[data-test=PlacePickerInputPlace-close]", state="hidden")
 
-    # 2. Hit the left datepicker month button and store the month value into a variable
-    departure_month_button = page.locator("[data-test=DatepickerMonthButton]").nth(0)
-    departure_month_button.click()
-    departure_month_value = departure_month_button.inner_text().split()[0]
+    # 1.3. Type in `Brno` to the `from` field
+    page.fill("[data-test=PlacePickerInput-origin] [data-test=SearchField-input]", "Brno")
 
-    # 3. Hit the right datepicker month button and store the month value into a variable
-    return_month_button = page.locator("[data-test=DatepickerMonthButton]").nth(1)
-    return_month_button.click()
-    return_month_value = return_month_button.inner_text().split()[0]
+    # 1.4. Select the 1st result from the dropdown
+    page.click("[data-test=PlacePickerRow-wrapper]:has-text('Brno, Czechia')")
 
-    # 4. Verify the Departure field contains a string in the following format: Wkd D Mmm – Wkd DD Mmm, where:
-    # * Wkd are the first 3 characters of a weekday, e.g., Mon or Wed (can be hardcoded)
-    # * D is a numeric value of the first day of the month, i.e., 1 (can be hardcoded)
-    # * DD is a numeric value of the last day of the month, e.g., 28, 29, 30, or 31 (can be hardcoded)
-    # * Mmm are the first 3 characters of name of the month, e.g., Mar or Jun - for each date field it should correspond with the value extracted from steps 3. and 4., respectively
-    weekdays = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]
+    # 1.5. Type in `Bucharest` to the `to` field
+    page.fill("[data-test=PlacePickerInput-destination] [data-test=SearchField-input]", "Bucharest")
 
-    departure_field_contents = page.locator(
-        "[class*=Inputsstyled__InputWrap]:has([data-navigation=DepartureRangeHeading]) [data-test=DateValue]"
-    ).inner_text().split()
+    # 1.6. Select the 1st result from the dropdown
+    page.click("[data-test=PlacePickerRow-wrapper]:has-text('Bucharest, Romania')")
 
-    departure_weekday_1 = departure_field_contents[0]
-    departure_day_number_1 = int(departure_field_contents[1])
-    departure_month_1 = departure_field_contents[2]
-    departure_weekday_2 = departure_field_contents[4]
-    departure_day_number_2 = int(departure_field_contents[5])
-    departure_month_2 = departure_field_contents[6]
+    # 1.7. Uncheck the `Booking` checkbox
+    page.click("[class*=BookingcomSwitchstyled] [class*=Checkbox]")
 
-    assert departure_weekday_1 in weekdays
-    assert departure_weekday_2 in weekdays
-    assert departure_day_number_1 in range(1, 32)
-    assert departure_day_number_2 in range(1, 32)
-    assert departure_month_1 == departure_month_value[:3]
-    assert departure_month_2 == departure_month_value[:3]
+    # 1.8. Hit the `Search` button
+    page.click("[data-test=LandingSearchButton]")
 
-    # 5. Verify the Departure field contains a string in the following format: Wkd D Mmm – Wkd DD Mmm
-    return_field_contents = page.locator(
-        "[class*=Inputsstyled__InputWrap]:has([data-test=ReturnRangeHeading]) [data-test=DateValue]"
-    ).inner_text().split()
+    # 1.9. Available connections should be displayed
+    page.wait_for_selector("[class*=ResultListstyled__ResultListWrapper]", timeout=10000)
+    page.wait_for_selector("[data-test=ResultCardWrapper]", state="visible")
 
-    return_weekday_1 = return_field_contents[0]
-    return_day_number_1 = int(return_field_contents[1])
-    return_month_1 = return_field_contents[2]
-    return_weekday_2 = return_field_contents[4]
-    return_day_number_2 = int(return_field_contents[5])
-    return_month_2 = return_field_contents[6]
+    # 2. Check the Train checkbox in the Transport left-hand section of the results
+    page.click("[class*=FilterWrapper]:has([data-test=TransportOptionCheckbox-train])")
+    page.wait_for_selector("[data-test=ResultList] [class*=LoadingProvidersstyled]", state="visible")
+    page.wait_for_selector("[data-test=ResultList] [class*=LoadingProvidersstyled]", state="hidden")
 
-    assert return_weekday_1 in weekdays
-    assert return_weekday_2 in weekdays
-    assert return_day_number_1 in range(1, 32)
-    assert return_day_number_2 in range(1, 32)
-    assert return_month_1 == return_month_value[:3]
-    assert return_month_2 == return_month_value[:3]
+    # 3. Select the Cheapest sorting option from the sorting panel
+    page.click("[data-test=SortBy-price]")
+    page.wait_for_selector("[data-test=ResultList] [class*=LoadingProvidersstyled]", state="visible")
+    page.wait_for_selector("[data-test=ResultList] [class*=LoadingProvidersstyled]", state="hidden")
+
+    # 4. Store the price value of the first few (e.g., 5) results into a (Python) list
+    result_values_list = []
+    for i in range(5):
+        nth_result_with_currency_code = (
+            page.locator("[data-test=ResultCardPrice][class*=ResultCardstyled__PriceText]")
+            .nth(i)
+            .inner_text()
+            .split()[0]
+        )
+        nth_result_value = int(nth_result_with_currency_code.replace(",", ""))
+        result_values_list.append(nth_result_value)
+
+    # 5. Verify the first result is either the cheapest of the stored values, or the same as the rest of them
+    assert result_values_list[0] <= result_values_list[1]
+
+    # (6. variation: on step 3. select the Fastest sorting option; on step 5. identify the cheapest of the stored
+    # results [i.e., identify the cheapest of the several fastest connections)
+    # 6.1. Select the Fastest sorting option
+    page.click("[data-test=SortBy-duration]")
+    page.wait_for_selector("[data-test=ResultList] [class*=LoadingProvidersstyled]", state="visible")
+    page.wait_for_selector("[data-test=ResultList] [class*=LoadingProvidersstyled]", state="hidden")
+
+    # 6.2. Identify the cheapest of the several fastest connections
+    result_values_list = []
+    for i in range(5):
+        nth_result_with_currency_code = (
+            page.locator("[data-test=ResultCardPrice][class*=ResultCardstyled__PriceText]")
+            .nth(i)
+            .inner_text()
+            .split()[0]
+        )
+        nth_result_value = int(nth_result_with_currency_code.replace(",", ""))
+        result_values_list.append(nth_result_value)
+
+    print(f"The cheapest of the several fastest connections costs {min(result_values_list)} CZK")
