+++ exercises/solutions/solution_08.py
@@ -1,46 +1,70 @@
-# Sidebar actions (expanding options, verifying visibility of items)
-def test_sidebar_actions_work_as_expected(page):
-    # 1. On the Kiwi.com website hit the right sidebar hamburger button
+# Localized currency is retained in Passenger details
+def test_localized_currency_is_retained_in_passenger_details(page):
+    # 1. Hit the 🇬🇧 CZK button in the navigation bar at the top of the search page
     # 1.1. Open the kiwi.com website (wait for page to load)
     page.goto("https://www.kiwi.com/en/")
     page.click("[data-test='CookiesPopup-Accept']")
     assert page.is_visible("text=Book cheap flights other sites simply can’t find.")
 
-    # 1.2. Hit the right sidebar hamburger button
-    page.click("[data-test=NavBar-SideNav-Open]")
-    page.wait_for_selector("[data-test=NavBar-SideNav][aria-hidden=false]", state="visible")
+    # 1.2 Open the localization settings and wait for the localization modal to be displayed
+    page.click("[data-test=RegionalSettingsButton]")
+    assert page.is_visible("[data-test=RegionalSettingsModal]")
 
-    # 2. Verify a sidebar with the Manage your trips, set up price alerts, use Kiwi.com Credit, and get personalized support. text appears
-    expected_sidebar_text = "Manage your trips, set up price alerts, use Kiwi.com Credit, and get personalized support."
-    current_sidebar_text = page.locator("[data-test=NavBar-SideNav] [class*=Text]").first.inner_text()
-    assert expected_sidebar_text == current_sidebar_text
+    # 2. Set the currency in the Language and currency modal to EUR
+    currency_code = "EUR"
+    page.select_option("[data-test=CurrencySelect]", value=f"{currency_code.lower()}")
 
-    # 3. Hit the Discover button
-    page.click("[data-test=NavBar-SideNav][aria-hidden=false] [role=button]:has-text('Discover') [aria-expanded=false]")
+    # 3. Hit the Save & continue button
+    page.click("[data-test=SubmitRegionalSettingsButton]")
 
-    # 4. Verify the Discover button expands into a dropdown/slide of items
-    assert page.is_visible(
-        "[data-test=NavBar-SideNav][aria-hidden=false] [role=button]:has-text('Discover') [aria-expanded=true]"
-    )
+    # 4. Search for connections between any two cities (while un-checking the Booking.com checkbox, as in previous scenarios)
+    # 4.1. Clear the `from` location (here with a stabilization to ensure the place-chip is always removed)
+    page.click("[data-test=PlacePickerInputPlace-close]")
+    if page.is_visible("[data-test=PlacePickerInputPlace-close]"):
+        page.click("[data-test=PlacePickerInputPlace-close]")
+    page.wait_for_selector("[data-test=PlacePickerInputPlace-close]", state="hidden")
 
-    # 5. Verify the Subscribe to newsletter button is displayed
-    assert page.is_visible("[aria-hidden=false] [class*=TextLink]:has-text('Subscribe to newsletter')")
+    # 4.2. Type in `Brno` to the `from` field
+    page.fill("[data-test=PlacePickerInput-origin] [data-test=SearchField-input]", "Brno")
 
-    # 6. Verify the Stories button is displayed
-    assert page.is_visible("[aria-hidden=false] [class*=TextLink]:has-text('Stories')")
+    # 4.3. Select the 1st result from the dropdown
+    page.click("[data-test=PlacePickerRow-wrapper]")
 
-    # 7. Hit the Subscribe to newsletter button
-    page.click("[aria-hidden=false] [class*=TextLink]:has-text('Subscribe to newsletter')")
+    # 4.4. Type in `Bucharest` to the `to` field
+    page.fill("[data-test=PlacePickerInput-destination] [data-test=SearchField-input]", "Bucharest")
 
-    # 8. Verify the sidebar disappears
-    page.wait_for_selector("[data-test=NavBar-SideNav][aria-hidden=false]", state="hidden")
+    # 4.5. Select the 1st result from the dropdown
+    page.click("[data-test=PlacePickerRow-wrapper]")
 
-    # 9. Verify a modal with the Subscribe to the Kiwi.com newsletter heading is displayed
-    page.wait_for_selector("[class*=Modal__ModalBody]", state="visible")
-    assert page.is_visible(
-        "[class*=Modal__ModalBody] [class*=Heading]:has-text('Subscribe to the Kiwi.com newsletter')"
-    )
+    # 4.6. Uncheck the `Booking` checkbox
+    page.click("[class*=BookingcomSwitchstyled] [class*=Checkbox]")
 
-    # 10. Verify the modal can be closed by hitting the cross button in its top right corner
-    page.click("[data-test=ModalCloseButton]")
-    page.wait_for_selector("[class*=Modal__ModalBody]", state="hidden")
+    # 4.7. Hit the `Search` button
+    page.click("[data-test=LandingSearchButton]")
+
+    # 4.8. Available connections should be displayed
+    page.wait_for_selector("[class*=ResultListstyled__ResultListWrapper]", timeout=10000)
+    page.wait_for_selector("[data-test=ResultCardWrapper]", state="visible")
+
+    # 5. Store the price value of the first result
+    first_result_without_currency_code = page.locator("[data-test=ResultCardPrice]").first.inner_text().split()[0]
+    first_result_value = int(first_result_without_currency_code)
+
+    # 6. Hit the Select button of the first result
+    page.click("[data-test=BookingButton]")
+    page.wait_for_selector("[data-test=MagicLogin]", state="visible")
+
+    # 7. In the Want to sign first? modal hit the Continue as a guest link
+    page.click("[data-test=MagicLogin-GuestTextLink]")
+    page.wait_for_selector("[data-test=ResultCardWrapper]", state="hidden")
+    page.wait_for_selector("[data-test=Reservation-content]", state="visible")
+    page.wait_for_selector("[data-test=Breadcrumbs-step-PASSENGER] [aria-current=step]", state="visible")
+
+    # 8. Verify the Total (EUR) price value corresponds with the one stored on step 5.
+    total_price_with_currency_code = page.locator("[class*=ReservationBillTotal] [class*=Price]").inner_text()
+    total_price_value = int(total_price_with_currency_code.split()[0])
+    assert first_result_value == total_price_value
+
+    # (9. variation: verify the currency code selected on step 1 is displayed next to Total in the reservation bill)
+    total_currency_code = page.locator("[data-test=ReservationBillBoxTotal]").inner_text().split("(")[1].strip(")")
+    assert currency_code == total_currency_code
