+++ exercises/tasks/task_03.py
@@ -1,35 +1,25 @@
-# Sorting panel actions can be used for sorting the search results
-def test_sorting_panel_actions_can_be_used_for_sorting_the_search_results(page):
-    # 1. Steps 1-9. from the Searching for a connection displays results scenario, but with the From location
-    # set to Brno and the To location set to Bucharest
+# Sidebar actions (expanding options, verifying visibility of items)
+def test_sidebar_actions_work_as_expected(page):
+    # 1. On the Kiwi.com website hit the right sidebar hamburger button
     # 1.1. Open the kiwi.com website (wait for page to load)
     pass
-    # 1.2. Clear the `from` location
 
-    # 1.3. Type in `Brno` to the `from` field
+    # 1.2. Hit the right sidebar hamburger button
 
-    # 1.4. Select the 1st result from the dropdown
+    # 2. Verify a sidebar with the Manage your trips, set up price alerts, use Kiwi.com Credit, and get personalized support. text appears
 
-    # 1.5. Type in `Bucharest` to the `to` field
+    # 3. Hit the Discover button
 
-    # 1.6. Select the 1st result from the dropdown
+    # 4. Verify the Discover button expands into a dropdown/slide of items
 
-    # 1.7. Uncheck the `Booking` checkbox
+    # 5. Verify the Subscribe to newsletter button is displayed
 
-    # 1.8. Hit the `Search` button
+    # 6. Verify the Stories button is displayed
 
-    # 1.9. Available connections should be displayed
+    # 7. Hit the Subscribe to newsletter button
 
-    # 2. Check the Train checkbox in the Transport left-hand section of the results
+    # 8. Verify the sidebar disappears
 
-    # 3. Select the Cheapest sorting option from the sorting panel
+    # 9. Verify a modal with the Subscribe to the Kiwi.com newsletter heading is displayed
 
-    # 4. Store the price value of the first few (e.g., 5) results into a (Python) list
-
-    # 5. Verify the first result is either the cheapest of the stored values, or the same as the rest of them
-
-    # (6. variation: on step 3. select the Fastest sorting option; on step 5. identify the cheapest of the stored
-    # results [i.e., identify the cheapest of the several fastest connections)
-    # 6.1. Select the Fastest sorting option
-
-    # 6.2. Identify the cheapest of the several fastest connections
+    # 10. Verify the modal can be closed by hitting the cross button in its top right corner
