+++ exercises/tasks/task_07.py
@@ -1,19 +1,35 @@
-# Calendar buttons (hit month and expect it in the date field)
-def test_hitting_calendar_buttons_is_reflected_by_date_field(page):
-    # 1. On the Kiwi.com website hit the Departure date field
+# Sorting panel actions can be used for sorting the search results
+def test_sorting_panel_actions_can_be_used_for_sorting_the_search_results(page):
+    # 1. Steps 1-9. from the Searching for a connection displays results scenario, but with the From location
+    # set to Brno and the To location set to Bucharest
     # 1.1. Open the kiwi.com website (wait for page to load)
     pass
+    # 1.2. Clear the `from` location
 
-    # 1.2. Hit the Departure date field
+    # 1.3. Type in `Brno` to the `from` field
 
-    # 2. Hit the left datepicker month button and store the month value into a variable
+    # 1.4. Select the 1st result from the dropdown
 
-    # 3. Hit the right datepicker month button and store the month value into a variable
+    # 1.5. Type in `Bucharest` to the `to` field
 
-    # 4. Verify the Departure field contains a string in the following format: Wkd D Mmm – Wkd DD Mmm, where:
-    # * Wkd are the first 3 characters of a weekday, e.g., Mon or Wed (can be hardcoded)
-    # * D is a numeric value of the first day of the month, i.e., 1 (can be hardcoded)
-    # * DD is a numeric value of the last day of the month, e.g., 28, 29, 30, or 31 (can be hardcoded)
-    # * Mmm are the first 3 characters of name of the month, e.g., Mar or Jun - for each date field it should correspond with the value extracted from steps 3. and 4., respectively
+    # 1.6. Select the 1st result from the dropdown
 
-    # 5. Verify the Departure field contains a string in the following format: Wkd D Mmm – Wkd DD Mmm
+    # 1.7. Uncheck the `Booking` checkbox
+
+    # 1.8. Hit the `Search` button
+
+    # 1.9. Available connections should be displayed
+
+    # 2. Check the Train checkbox in the Transport left-hand section of the results
+
+    # 3. Select the Cheapest sorting option from the sorting panel
+
+    # 4. Store the price value of the first few (e.g., 5) results into a (Python) list
+
+    # 5. Verify the first result is either the cheapest of the stored values, or the same as the rest of them
+
+    # (6. variation: on step 3. select the Fastest sorting option; on step 5. identify the cheapest of the stored
+    # results [i.e., identify the cheapest of the several fastest connections)
+    # 6.1. Select the Fastest sorting option
+
+    # 6.2. Identify the cheapest of the several fastest connections
